move-spec-decoding-meta-move: Move _calculate_spec_decoding_metrics to managers/tokenizer_manager_components/spec_decoding_meta.py as fill_spec_decoding_meta
Physical move only:
  - Cut staticmethod from TM
  - Drop @staticmethod, dedent
  - Rename _calculate_spec_decoding_metrics -> fill_spec_decoding_meta
  - Write managers/tokenizer_manager_components/spec_decoding_meta.py
  - Caller: TokenizerManager._calculate_spec_decoding_metrics(...) ->
    spec_decoding_meta.fill_spec_decoding_meta(...)

Refactor chain ID: move-spec-decoding-meta-move
